backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pdfplumber==0.10.3
spacy==3.7.2
scikit-learn==1.3.2
python-dotenv==1.0.0
cors==1.0.1
